backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
docker==7.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
psutil==5.9.8
aiofiles==23.2.1
structlog==24.1.0
